Fix Mask dtype
	1. add_mask always propoagates previous_mask.dtype
	2. merge_mask expects same dtype
	3. create_empty_mask should be called with the correct dtype
	   (e.g. existing tensor's dtype)

	Tool: Human

+++ sparse_attention_hub/sparse_attention/research_attention/base.py
@@ -82,7 +82,7 @@ class ResearchAttention(SparseAttention):
             queries.shape[2],
             keys.shape[2],
         )
-        sparse_attention_mask = Mask.create_empty_mask(mask_shape)
+        sparse_attention_mask = Mask.create_empty_mask(mask_shape, dtype=queries.dtype)
 
         # Apply all maskers sequentially, each one on the output of the previous one
         for masker in self.maskers:
